file 2

+++ pointers_arrays_strings/2-strncpy.c
@@ -1,9 +1,9 @@
 #include"main.h"
 
 
-char *_strncpy(char *dest, const char *src, size_t n)
+char *_strncpy(char *dest, const char *src, int n)
 {
-    size_t i;
+    int i;
 
     for (i = 0; i < n && src[i] != '\0'; i++) {
         dest[i] = src[i];
